Refactored code for `getImageBase64`
