+++ app/server/getImageBase64/index.ts
@@ -1,0 +1,1 @@
+export * from "./getImageBase64";
